\name{phcfM-package}
\alias{phcfM-package}
\alias{phcfM}
\docType{package}
\title{
  Modelling anthropogenic deforestation in tropical forest
}

\description{phcfM is an R package for modelling anthropogenic
deforestation. It was initially developed to obtain REDD baseline
scenarios of deforestation for the "programme holistique de conservation
des forets a Madagascar" (from which the package was named after).}

\details{
\tabular{ll}{
  Package: \tab phcfM\cr
  Type: \tab Package\cr
  Version: \tab 1.0\cr
  Date: \tab 2012-03-06\cr
  License: \tab GPL \cr
  LazyLoad: \tab yes\cr
}
}

\author{
  Ghislain Vieilledent <ghislain.vieilledent@cirad.fr>
}

\keyword{tropical forest}
\keyword{deforestation}
\keyword{anthropogenic pressure}
